Final touches

+++ src/model/Cost.h
@@ -3,34 +3,35 @@
 #include <map>
 #include <vector>
 #include <compare>
+#include <cstdint>
 #include "JsonUtils.h"
 #include "CoreExport.h"
 
 class CORE_API Cost {
 private:
-	int m_coinCost;
-	std::map<ResourceType, int> m_resourceCosts;
+	uint8_t m_coinCost;
+	std::map<ResourceType, uint8_t> m_resourceCosts;
 public:
 	Cost();
-	Cost(int coins);
-	Cost(int coins, const std::map<ResourceType, int>& resources);
+	Cost(uint8_t coins);
+	Cost(uint8_t coins, const std::map<ResourceType, uint8_t>& resources);
 	Cost(const Cost&) = default;
 	Cost& operator=(const Cost&) = default;
 	Cost(Cost&&) = default;
 	Cost& operator=(Cost&&) = default;
 	~Cost() = default;
 
-	Cost& withCoinCost(int coins);
-	Cost& withResourceCost(ResourceType type, int amount);
+	Cost& withCoinCost(uint8_t coins);
+	Cost& withResourceCost(ResourceType type, uint8_t amount);
 
-	int getCoinCost() const;
-	const std::map<ResourceType, int>& getResourceCosts() const;
-	int getResourceCost(ResourceType type) const;
+	uint8_t getCoinCost() const;
+	const std::map<ResourceType, uint8_t>& getResourceCosts() const;
+	uint8_t getResourceCost(ResourceType type) const;
 	bool isFree() const;
 
 	bool hasResourceCost() const;
 	bool hasCoinCost() const;
-	int  getTotalResourceCount() const;
+	int getTotalResourceCount() const;
 
 	std::string toString() const;
 	std::string toShortString() const;
@@ -50,7 +51,7 @@ public:
 	Cost& operator-=(const Cost& discount);
 
 	std::strong_ordering operator<=>(const Cost& other) const;
-    bool operator==(const Cost& other) const;
+	bool operator==(const Cost& other) const;
 
 	friend std::ostream& operator<<(std::ostream& os, const Cost& cost);
